requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.20.0
deepspeed>=0.12.0

# Data processing
numpy>=1.24.0
pillow>=9.5.0
opencv-python>=4.8.0
decord>=0.6.0

# Training
trl>=0.7.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Evaluation
nltk>=3.8.0
rouge-score>=0.1.2
scikit-learn>=1.3.0

# SAM3 (will be installed separately)
# ultralytics>=8.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0

# Audio processing
av>=10.0.0
librosa>=0.10.0

# Video processing
ffmpeg-python>=0.2.0
